% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descfarspkg_functions.R
\name{fars_map_state}
\alias{fars_map_state}
\title{fars_map_state}
\usage{
fars_map_state(state.num, year)
}
\arguments{
\item{state.num}{A numeric vector representing the number of the American state: 1-51}

\item{year}{A numeric vector representing the years: 2013-2015}
}
\value{
a map of American state and points of fatalities
}
\description{
This function plots the map of pointwise car fatalities in given year at a given American state.
}
\examples{
\dontrun{fars_map_state(10,2015)}

}
