% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descfarspkg_functions.R
\name{fars_read}
\alias{fars_read}
\title{fars_read}
\source{
\url{http://www.nhtsa.gov/Data/Fatality-Analysis-Reporting-System-(FARS)}
}
\usage{
fars_read(filename)
}
\arguments{
\item{filename}{A character string with the name of the file to read}
}
\value{
This function returns a tbl_df dataframe object

An error message if the data does not exist or a data frame with data readed from the csv input file.
}
\description{
This function reads file from .csv input file.
}
\examples{
\dontrun{fars_read("accident_2015.csv.bz2")}

}
