% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descfarspkg_functions.R
\name{fars_read_years}
\alias{fars_read_years}
\title{fars_read_years}
\usage{
fars_read_years(years)
}
\arguments{
\item{years}{A numeric vector representing the years}
}
\value{
a tibble data-matrix of month and year and fatalities in that month
}
\description{
This function reads data per input year and returns a tibble data matrix of 2 columns (Month,year).
}
\examples{
\dontrun{fars_read_years(c(2014,2015))}

}
