% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descfarspkg_functions.R
\name{fars_summarize_years}
\alias{fars_summarize_years}
\title{fars_summarize_years}
\usage{
fars_summarize_years(years)
}
\arguments{
\item{years}{A numeric vector representing the years}
}
\value{
a tibble data-matrix of 12 by 2 (counts per month)
}
\description{
This function gets the year in the datset and returns a tibble datamatrix of two columns showing the counts of fatalities by month
}
\examples{
\dontrun{fars_summarize_years(c(2014,2015))}

}
